refactor(tests): deduplicate fusion helpers
